requirements/default.txt
docker-py == 1.8.1
pyyaml    == 3.11
funcy     == 1.7.1
